treewide: Fix float to double promotion error
Implicit promotion from float to double type now results in a
compilation warning, treated by CI as an error, hence use
explicit type cast wherever applicable.

+++ applications/asset_tracker_v2/src/modules/location_module.c
@@ -395,10 +395,13 @@ void location_event_handler(const struct location_event_data *event_data)
 		LOG_DBG("  method: %s", location_method_str(event_data->method));
 		LOG_DBG("  latitude: %.06f", event_data->location.latitude);
 		LOG_DBG("  longitude: %.06f", event_data->location.longitude);
-		LOG_DBG("  accuracy: %.01f m", event_data->location.accuracy);
-		LOG_DBG("  altitude: %.01f m", event_data->location.details.gnss.pvt_data.altitude);
-		LOG_DBG("  speed: %.01f m", event_data->location.details.gnss.pvt_data.speed);
-		LOG_DBG("  heading: %.01f deg", event_data->location.details.gnss.pvt_data.heading);
+		LOG_DBG("  accuracy: %.01f m", (double)event_data->location.accuracy);
+		LOG_DBG("  altitude: %.01f m",
+			(double)event_data->location.details.gnss.pvt_data.altitude);
+		LOG_DBG("  speed: %.01f m",
+			(double)event_data->location.details.gnss.pvt_data.speed);
+		LOG_DBG("  heading: %.01f deg",
+			(double)event_data->location.details.gnss.pvt_data.heading);
 
 		if (event_data->location.datetime.valid) {
 			LOG_DBG("  date: %04d-%02d-%02d",
