+++ ext/iperf3/iperf_api.c
@@ -3027,7 +3027,7 @@ static cJSON
                         str = (char *) calloc(sizeof(char), hsize + 1);   /* +1 for trailing null */
 			if (str == NULL) {
                         	iperf_printf(
-					test, 
+					test,
 					"ERROR: No memory for parsing json results of len %d\n",
 					hsize);
 				i_errno = IERECVRESULTS;
@@ -5269,7 +5269,7 @@ static void print_interval_results(struct iperf_test *test,
 	}
 
 	unit_snprintf(ubuf, UNIT_LEN, (double)(irp->bytes_transferred), 'A');
-	if (irp->interval_duration > 0.0) {
+	if ((double)irp->interval_duration > 0.0) {
 		bandwidth = (double)irp->bytes_transferred /
 			    (double)irp->interval_duration;
 	} else {
